Split session module in directory
