+++ services/orchest-api/app/app/core/sessions/__init__.py
@@ -1,0 +1,4 @@
+from app.core.sessions._core import (  # noqa
+    InteractiveSession,
+    launch_noninteractive_session,
+)
